refactor: 중복 QuestionViewModel, QuestionUiState 문제 해결(#136)
