+++ feature/quiz/src/main/java/kr/boostcamp_2024/course/quiz/viewmodel/UserQuestionViewModel.kt
@@ -1,4 +1,4 @@
-package kr.boostcamp_2024.course.quiz.presentation.viewmodel
+package kr.boostcamp_2024.course.quiz.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.SavedStateHandle
@@ -7,7 +7,10 @@ import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.onStart
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kr.boostcamp_2024.course.domain.model.BaseQuiz
@@ -21,21 +24,22 @@ import kr.boostcamp_2024.course.quiz.R
 import kr.boostcamp_2024.course.quiz.navigation.QuestionRoute
 import javax.inject.Inject
 
-data class QuestionUiState(
+data class UserQuestionUiState(
     val quiz: BaseQuiz? = null,
     val questions: List<Question> = emptyList(),
     val isSubmitting: Boolean = false,
     val currentPage: Int = 0,
     val selectedIndexList: List<Int> = emptyList(),
-    val countDownTime: Int = 20 * 60,
+    val submittedIndexList: List<Int> = emptyList(),
     val isLoading: Boolean = false,
     val errorMessageId: Int? = null,
     val currentUserId: String? = null,
     val userOmrId: String? = null,
+    val isSubmitted: Boolean = false,
 )
 
 @HiltViewModel
-class QuestionViewModel @Inject constructor(
+class UserQuestionViewModel @Inject constructor(
     private val authRepository: AuthRepository,
     private val questionRepository: QuestionRepository,
     private val quizRepository: QuizRepository,
@@ -44,10 +48,20 @@ class QuestionViewModel @Inject constructor(
 ) : ViewModel() {
     private val quizId = savedStateHandle.toRoute<QuestionRoute>().quizId
 
-    private val _uiState: MutableStateFlow<QuestionUiState> = MutableStateFlow(QuestionUiState())
-    val uiState: StateFlow<QuestionUiState> = _uiState
+    private val _uiState: MutableStateFlow<UserQuestionUiState> = MutableStateFlow(
+        UserQuestionUiState(),
+    )
+    val uiState: StateFlow<UserQuestionUiState> = _uiState
+        .onStart {
+            initial()
+            updatePageAndSubmitByOwner()
+        }.stateIn(
+            viewModelScope,
+            SharingStarted.WhileSubscribed(5000L),
+            UserQuestionUiState(),
+        )
 
-    fun initial() {
+    private fun initial() {
         viewModelScope.launch {
             loadCurrentUserId()
 
@@ -61,23 +75,6 @@ class QuestionViewModel @Inject constructor(
                     Log.e("QuestionViewModel", "퀴즈 로드 실패", it)
                     _uiState.update { it.copy(isLoading = false, errorMessageId = R.string.err_load_quiz) }
                 }
-            updateTimer()
-        }
-    }
-
-    private fun loadCurrentUserId() {
-        viewModelScope.launch {
-
-            authRepository.getUserKey()
-                .onSuccess { currentUser ->
-
-                    _uiState.update { it.copy(currentUserId = currentUser) }
-                }
-                .onFailure {
-                    Log.e("MainViewModel", "Failed to load current user", it)
-                    _uiState.update { it.copy(errorMessageId = R.string.err_load_current_user) }
-                }
-
         }
     }
 
@@ -92,6 +89,7 @@ class QuestionViewModel @Inject constructor(
                         currentState.copy(
                             questions = questions,
                             selectedIndexList = List<Int>(questions.size) { -1 },
+                            submittedIndexList = List<Int>(questions.size) { -1 },
                             isLoading = false,
                         )
                     }
@@ -104,6 +102,19 @@ class QuestionViewModel @Inject constructor(
                             errorMessageId = R.string.err_load_questions,
                         )
                     }
+                }
+        }
+    }
+
+    private fun loadCurrentUserId() {
+        viewModelScope.launch {
+            authRepository.getUserKey()
+                .onSuccess { currentUser ->
+                    _uiState.update { it.copy(currentUserId = currentUser) }
+                }
+                .onFailure {
+                    Log.e("MainViewModel", "Failed to load current user", it)
+                    _uiState.update { it.copy(errorMessageId = R.string.err_load_current_user) }
                 }
         }
     }
@@ -124,25 +135,48 @@ class QuestionViewModel @Inject constructor(
         }
     }
 
-    fun nextPage() {
-        _uiState.update { currentState ->
-            currentState.copy(currentPage = currentState.currentPage + 1)
-        }
-    }
-
-    fun previousPage() {
-        _uiState.update { currentState ->
-            currentState.copy(currentPage = currentState.currentPage - 1)
-        }
-    }
-
-    fun updateTimer() {
+    private fun updatePageAndSubmitByOwner() {
         viewModelScope.launch {
-            while (_uiState.value.countDownTime > 0) {
-                kotlinx.coroutines.delay(1000L)
-                _uiState.update { currentState ->
-                    currentState.copy(countDownTime = currentState.countDownTime - 1)
+            quizRepository.observeCurrentPage(quizId)
+                .collect { result ->
+                    result
+                        .onSuccess { newCurrentPage ->
+                            newCurrentPage?.let {
+
+                                if (it == _uiState.value.questions.size && it != 0) { // questions.size를 늦게 불러와서 오류뜸
+                                    submitAnswers()
+                                    _uiState.update {
+                                        it.copy(currentPage = 0)
+                                    }
+                                } else {
+                                    _uiState.update {
+                                        it.copy(
+                                            currentPage = newCurrentPage,
+                                            isSubmitted = false,
+                                        )
+                                    }
+                                }
+                            }
+                        }.onFailure {
+                            Log.e("UserQuestionViewModel", "페이지 로드 실패", it)
+                            _uiState.update { it.copy(errorMessageId = R.string.err_load_current_page) }
+                        }
                 }
+        }
+    }
+
+    fun submitQuestion(questionId: String) {
+        viewModelScope.launch {
+            val result = questionRepository.updateCurrentSubmit(questionId)
+            val updatedList = _uiState.value.submittedIndexList.toMutableList().apply {
+                this[_uiState.value.currentPage] =
+                    _uiState.value.selectedIndexList[_uiState.value.currentPage]
+            }
+            result.onSuccess {
+                _uiState.update { it.copy(isSubmitted = true, submittedIndexList = updatedList) }
+            }.onFailure {
+                Log.e("UserQuestionViewModel", "제출 실패 ", it)
+                _uiState.update { it.copy(errorMessageId = R.string.err_submit_quetion) }
             }
         }
     }
@@ -155,7 +189,7 @@ class QuestionViewModel @Inject constructor(
                 val userOmrCreationInfo = UserOmrCreationInfo(
                     userId = userId,
                     quizId = quizId,
-                    answers = _uiState.value.selectedIndexList.map { it },
+                    answers = _uiState.value.submittedIndexList.map { it },
                 )
                 userOmrRepository.submitQuiz(userOmrCreationInfo)
                     .onSuccess { userOmrId ->
